Mildly bugged bucket system in. Added a rectangle class to define a region to pass into the hash to get entities from. Made UI_Panel inherit from it. Made Vector2 comparison const and a hash function for the map to use.

+++ hTech/UI_Panel.cpp
@@ -5,7 +5,7 @@ UI_Panel::UI_Panel()
 {
 }
 
-UI_Panel::UI_Panel(unsigned short posX, unsigned short posY, unsigned short width, unsigned short height, Colour textColour, Colour bgColour)
+UI_Panel::UI_Panel(int posX, int posY, int width, int height, Colour textColour, Colour bgColour)
 {
 	this->X = posX;
 	this->Y = posY;
